+++ packages/mds-policy-author-api/logger.ts
@@ -1,0 +1,3 @@
+import { createLogger } from '@mds-core/mds-logger'
+
+export const PolicyAuthorApiLogger = createLogger('mds-policy-author-api')